+++ docs/reference/changelog.rst
@@ -23,6 +23,7 @@ v5.5.0 - 2022-xx-xx
 
 - ``Changed`` Python patch bump + optimizations - by @goegol [`#1665 <https://github.com/dsmrreader/dsmr-reader/issues/1665>`_]
 - ``Changed`` Tariefnamen rechttrekken [`#1666 <https://github.com/dsmrreader/dsmr-reader/issues/1666>`_]
+- ``Changed`` Allow graph 'stack' option for live graphs [`#1420 <https://github.com/dsmrreader/dsmr-reader/issues/1420>`_]
 
 
 v5.4.0 - 2022-07-07